[nemo-qml-plugin-email] Move to utils.h the part-related functions from EmailAgent.
