+++ src/emailagent.h
@@ -90,8 +90,6 @@ public:
                                                           const QString &attachmentLocation,
                                                           QString *downloadPath = nullptr);
     double attachmentDownloadProgress(const QString &attachmentLocation);
-    QString attachmentName(const QMailMessagePart &part) const;
-    QString attachmentTitle(const QMailMessagePart &part) const;
     QString bodyPlainText(const QMailMessage &mailMsg) const;
     void cancelAction(quint64 actionId);
     quint64 downloadMessages(const QMailMessageIdList &messageIds, QMailRetrievalAction::RetrievalSpecification spec);
