+++ src/emailutils.h
@@ -22,6 +22,44 @@ inline bool isEmailPart(const QMailMessagePart &part)
     return false;
 }
 
+inline QString attachmentName(const QMailMessagePart &part)
+{
+    return part.displayName().remove('/');
+}
+
+inline QString attachmentTitle(const QMailMessagePart &part)
+{
+    if (isEmailPart(part)) {
+        if (part.contentAvailable())
+            return QMailMessage::fromRfc2822(part.body().data(QMailMessageBody::Decoded)).subject();
+
+        auto contentType = part.contentType();
+        QString name = contentType.isParameterEncoded("name")
+            ? QMailMessageHeaderField::decodeParameter(contentType.name()).trimmed()
+            : QMailMessageHeaderField::decodeContent(contentType.name()).trimmed();
+
+        // QMF plugin may append an extra .eml ending, remove both
+        for (int i = 0; name.endsWith(EML_EXTENSION) && i < 2; i++)
+            name.chop(4);
+
+        if (!name.isEmpty())
+            return name;
+
+        auto contentDisposition = part.contentDisposition();
+        name = contentDisposition.isParameterEncoded("filename")
+            ? QMailMessageHeaderField::decodeParameter(contentDisposition.filename()).trimmed()
+            : QMailMessageHeaderField::decodeContent(contentDisposition.filename()).trimmed();
+
+        if (name.endsWith(EML_EXTENSION))
+            name.chop(4);
+
+        if (!name.isEmpty())
+            return name;
+    }
+
+    return QString();
+}
+
 inline int attachmentSize(const QMailMessagePart &part)
 {
     if (part.contentDisposition().size() != -1) {
